fix: sqlx offline cache + fix ~35 compilation errors across all Rust services
- Génère le cache .sqlx/ offline via cargo sqlx prepare --workspace
- Ajoute ENV SQLX_OFFLINE=true + COPY .sqlx dans tous les Dockerfiles Rust
- Corrige les erreurs de compilation : bollard 0.17 (HostConfig aplati),
  sozu-command-lib 1.1.1 (RequestTcpFrontend), redis 0.25 (Value::Data),
  fred 8 (session store), never type fallback, types nullable/NOT NULL
- Ajoute migrations 202600010060 (users.is_active) et 202600010061
  (metadata, encryption_key_id, status, health_status, last_probed_at)
- Version 1.1.2

+++ apps/api/src/handlers/shared_terminal.rs
@@ -208,8 +208,8 @@ async fn run_owner_terminal(
             match msg {
                 Ok(Message::Binary(data)) => {
                     if data.len() == 5 && data[0] == 0x01 {
-                        let cols = u16::from_be_bytes([data[1], data[2]]) as u32;
-                        let rows = u16::from_be_bytes([data[3], data[4]]) as u32;
+                        let cols = u16::from_be_bytes([data[1], data[2]]);
+                        let rows = u16::from_be_bytes([data[3], data[4]]);
                         let _ = docker.resize_exec(&exec_id_clone, ResizeExecOptions { width: cols, height: rows }).await;
                     } else if input_writer.write(&data).await.is_err() {
                         break;
